Updated code so that the status code returned is an integer and also checking ther response in the unit test

+++ CodeBug.SendGridEmailer.Process/CodeBug.SendGridEmailer.Process/EmailService.cs
@@ -35,7 +35,7 @@ namespace CodeBug.SendGridEmailer.Process
 
                 var resposne = await httpClient.PostAsync(SendGirdV3EndPoint, new StringContent(messageBody, Encoding.UTF8, "application/json"));
                 
-                emailResponse.HttpStatusCode = resposne.StatusCode;
+                emailResponse.HttpStatusCode = (int)resposne.StatusCode;
 
                 if (!resposne.IsSuccessStatusCode)
                 {
